pyproject.toml
[tool.poetry]
name = "testcapitalcomplaywright"
version = "0.1.0"
description = ""
authors = ["Alexey Kurochkin <KurochkinAlexei@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
playwright = "^1.49.1"
pytest = "^8.3.4"
pytest-xdist = "^3.6.1"
allure-pytest = "^2.13.5"
pytest-html = "^4.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
